feat(game): refactor snake game logic and add new entity management
